move app into own library
